requirements.txt
playwright>=1.40.0
typer>=0.9.0
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
questionary>=2.0.0
